requirements.txt
mcp==0.9.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.0
httpx==0.25.2
python-dotenv==1.0.0
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.3
tree-sitter-typescript==0.20.3
PyGithub==1.59.1
atlassian-python-api==3.41.0
numpy==1.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
aiohttp==3.9.0
deimos-router
cohere
slack-bolt==1.18.1
certifi==2023.11.17
flask==3.0.0
